# specs/003-sso-auth-provider/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Keyles SSO Auth Provider API
  description: |
    OpenID Connect (OIDC) and OAuth 2.0 authentication provider for multi-tenant SaaS platform.
    
    **Authentication Flows**:
    - Authorization Code Flow with PKCE (RFC 7636)
    - Refresh Token Flow
    - Token Revocation (RFC 7009)
    
    **Security**:
    - RS256 JWT signing
    - Multi-tenant isolation
    - Role-based access control
    - Rate limiting (10 req/min per client)
    
  version: 1.0.0
  contact:
    name: Keyles Engineering
    email: eng@keyles.com
servers:
  - url: https://sso.keyles.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /oauth2/revoke:
    post:
      tags:
        - OAuth 2.0
      summary: Token Revocation
      description: |
        Revokes a refresh token (RFC 7009).
        Returns 200 OK even if token doesn't exist (security best practice).
      operationId: revokeToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
                - client_id
                - client_secret
              properties:
                token:
                  type: string
                  description: Token to revoke
                token_type_hint:
                  type: string
                  enum: [refresh_token]
                  description: Type of token (optional)
                client_id:
                  type: string
                  description: Client identifier
                client_secret:
                  type: string
                  description: Client secret
      responses:
        '200':
          description: Token revoked successfully (or didn't exist)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "revoked"
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
...
  /admin/roles:
    get:
      tags:
        - Admin - Roles
      summary: List Role Assignments
      description: Lists all user role assignments for the tenant
      operationId: listRoleAssignments
      security:
        - SessionAuth: []
      parameters:
        - name: client_id
          in: query
          schema:
            type: string
          description: Filter by client
        - name: user_id
          in: query
          schema:
            type: string
          description: Filter by user
      responses:
        '200':
          description: List of role assignments
          content:
            application/json:
              schema:
                type: object
                properties:
                  role_assignments:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoleAssignment'
    post:
      tags:
        - Admin - Roles
      summary: Assign Role
      description: Grants a user access to a client application with a specific role
      operationId: assignRole
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
            example:
              user_id: "user_12345"
              client_id: "abc123xyz"
              role: "user"
      responses:
        '201':
          description: Role assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignment'
        '400':
          description: Invalid request
  /admin/roles/{assignment_id}:
    delete:
      tags:
        - Admin - Roles
      summary: Revoke Role
      description: Revokes a user's access to a client application
      operationId: revokeRole
      security:
        - SessionAuth: []
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Role revoked successfully
        '404':
          description: Role assignment not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.0 access token
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie for admin endpoints
  schemas:
...
    RoleAssignment:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        client_id:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        granted_at:
          type: string
          format: date-time
        granted_by:
          type: string
    AssignRoleRequest:
      type: object
      required:
        - user_id
        - client_id
        - role
      properties:
        user_id:
          type: string
        client_id:
          type: string
        role:
          type: string
          enum: [admin, user, viewer]
    OAuth2Error:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - too_many_requests
        error_description:
          type: string
        error_uri:
          type: string
...